requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
python-dotenv==0.13.0
selenium==3.141.0